requirements.txt
Flask==0.10.1
Jinja2==2.7.3
python-jose==1.3.2
requests==1.1.0
